fix formatting

+++ src/main/groovy/cz/siret/prank/program/routines/ParamLooper.groovy
@@ -75,7 +75,7 @@ abstract class ParamLooper extends Routine {
 
     static String fmt(Object x) {
         if (x==null) return ""
-        sprintf "%8.4f", (double)x
+        sprintf "%8.4f", x
     }
 
 //===========================================================================================================//
